Checks for minimum required Python version 3.6
We don't go for the `required_python` setuptools parameter as it's just
filters the package that would appear or not. It means that previous
version would sill appear and even install on Python2. With this method
all version appear, but it throws a clear error message during the
install on version mismatch.